+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.30 branch
 ### 2.30.9-RC6
 * improved `\MeCms\Controller\SystemsController` class;
-* updated for me-tools 2.21.5.
+* updated for me-tools 2.21.5;
+* little fixes.
 
 ### 2.30.8-RC5
 * improved the `UserCommand`;