+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.9-RC6
-* improved `\MeCms\Controller\SystemsController` class;
+* improved `\MeCms\Controller\SystemsController` and `\MeCms\Form\ContactUsForm` classes;
 * updated for me-tools 2.21.5;
 * little fixes.